a little update to game.js
I learned about Phaser and decided to use your template for my first project. I noticed that you could "cheat" in the template by pausing the game and still clicking the hell out of the buttonDummy. You could also click in the GameOver state. This was my solution to that bug. First thing I did in Phaser. Thank you very much for the template, it was really helpful!